@@ -1,6 +1,5 @@
 package stakemate.use_case.PlaceOrderUseCase;
 
-
 import java.util.List;
 
 import stakemate.engine.BookOrder;
@@ -10,15 +9,35 @@ import stakemate.service.AccountService;
 
 /**
  * Orchestrates funds checks (optional) + matching engine.
+ * 
+ * Updated to support both In-Memory (legacy) and DB-backed (Postgres) modes.
  */
 public class PlaceOrderUseCase {
 
     private final MatchingEngine engine;
     private final AccountService accountService;
 
+    // --- DB Fields ---
+    private OrderRepository orderRepository;
+    private PositionRepository positionRepository;
+
+    /**
+     * Constructor for In-Memory mode.
+     */
     public PlaceOrderUseCase(final MatchingEngine engine, final AccountService accountService) {
         this.engine = engine;
         this.accountService = accountService;
+    }
+
+    /**
+     * Constructor for DB-backed mode.
+     */
+    public PlaceOrderUseCase(final MatchingEngine engine, final AccountService accountService,
+            final OrderRepository orderRepository, PositionRepository positionRepository) {
+        this.engine = engine;
+        this.accountService = accountService;
+        this.orderRepository = orderRepository;
+        this.positionRepository = positionRepository;
     }
 
     /**
@@ -37,17 +56,38 @@ public class PlaceOrderUseCase {
             return PlaceOrderResponse.fail("Insufficient funds");
         }
 
+        // --- DB Specific: Upfront cost deduction ---
+        if (orderRepository != null && req.price != null) {
+            double calcPrice = req.price;
+            double upfrontCost = calcPrice * req.quantity;
+
+            if (accountService instanceof stakemate.service.DbAccountService) {
+                stakemate.service.DbAccountService dbAcc = (stakemate.service.DbAccountService) accountService;
+                dbAcc.adjustBalance(req.userId, -upfrontCost);
+            }
+        }
+
         // create internal order (price == null => market)
         final BookOrder incoming = new BookOrder(req.userId, req.marketId, req.side, req.price, req.quantity);
         accountService.reserveForOrder(req.userId, incoming.getId(), estimateReservationAmount(incoming));
 
-        final List<Trade> trades = engine.placeOrder(incoming);
-        for (final Trade t : trades) {
-            // naive settlement: buyer pays price*size to seller
-            accountService.capture(t);
+        // --- DB Specific: Save order ---
+        if (orderRepository != null) {
+            orderRepository.save(incoming);
         }
 
-        final String msg = trades.isEmpty() ? "Order placed (no immediate trades)" : String.format("Executed %d trades", trades.size());
+        final List<Trade> trades = engine.placeOrder(incoming);
+
+        // --- In-Memory Specific: Naive settlement ---
+        if (orderRepository == null) {
+            for (final Trade t : trades) {
+                // naive settlement: buyer pays price*size to seller
+                accountService.capture(t);
+            }
+        }
+
+        final String msg = trades.isEmpty() ? "Order placed (no immediate trades)"
+                : String.format("Executed %d trades", trades.size());
         return PlaceOrderResponse.success(msg);
     }
 
@@ -64,5 +104,16 @@ public class PlaceOrderUseCase {
     public List<Trade> recentTrades() {
         return engine.getTrades();
     }
-}
 
+    /**
+     * Return all open orders (bids + asks) for a given user.
+     * This is used by the UI to populate the "Open Orders" table.
+     */
+    public List<BookOrder> openOrdersForUser(String userId) {
+        if (orderRepository != null) {
+            return orderRepository.findOpenOrdersForUser(userId);
+        }
+        // In-memory fallback (not implemented/needed for demo)
+        return java.util.Collections.emptyList();
+    }
+}
